Fixed remarks and refactor unit tests for UserDao

+++ chess-service/src/main/java/com/javamonkeys/dao/user/UserDao.java
@@ -25,18 +25,26 @@ public class UserDao implements IUserDao {
      */
     public User getUser(String email) throws UserNotFoundException {
 
-        HibernateUtil.begin();
-        Session session = HibernateUtil.getSession();
-        Query query = session.createQuery("from User where email = :email");
-        query.setParameter("email", email);
-        User currentUser = (User)query.uniqueResult();
-        HibernateUtil.commit();
+        if(email!=null) {
 
-        if(currentUser==null){
-            throw new UserNotFoundException("User with email: " + email + " not found.");
+            HibernateUtil.begin();
+            Session session = HibernateUtil.getSession();
+            Query query = session.createQuery("from User where email = :email");
+            query.setParameter("email", email);
+            User currentUser = (User)query.uniqueResult();
+            HibernateUtil.commit();
+            HibernateUtil.close();
+
+            if (currentUser == null) {
+                throw new UserNotFoundException("User with email: " + email + " was not found");
+            }
+
+            return currentUser;
+
+        }else{
+            throw new UserNotFoundException("User with null email was not found");
         }
 
-        return currentUser;
     }
 
     /**
@@ -47,30 +55,42 @@ public class UserDao implements IUserDao {
      * @return new user
      * @throws UserAlreadyExistException if user which this email already exist
      */
-    public User createUser(String email, String password, UserAccessGroup userAccessGroup) throws UserAlreadyExistException {
+    public User createUser(String email, String password, UserAccessGroup userAccessGroup) throws UserAlreadyExistException,
+            IllegalArgumentException {
 
-        try {
-            User currentUser = getUser(email);
-            if (currentUser != null) {
-                throw new UserAlreadyExistException("User with email: " + email + " is already exist.");
+        if(email==null) {
+            throw new IllegalArgumentException("email: " + email + " is not valid");
+        } else if (password==null) {
+            throw new IllegalArgumentException("password: " + password + " is not valid");
+        } else if (userAccessGroup==null) {
+            throw new IllegalArgumentException("userAccessGroup: " + userAccessGroup + " is not valid");
+        } else {
+
+            try {
+                User currentUser = getUser(email);
+                if (currentUser != null) {
+                    throw new UserAlreadyExistException("User with email: " + email + " is already exist.");
+                }
+            } catch (UserNotFoundException e) {
+                // ok - create a new user
             }
-        }catch (UserNotFoundException e){
-            // ok - create a new user
-        }
 
-        try {
+            try {
 
-            HibernateUtil.begin();
-            Session session = HibernateUtil.getSession();
-            User newUser = new User(email, password, userAccessGroup);
-            session.save(newUser);
-            HibernateUtil.commit();
+                HibernateUtil.begin();
+                Session session = HibernateUtil.getSession();
+                User newUser = new User(email, password, userAccessGroup);
+                session.save(newUser);
+                HibernateUtil.commit();
+                HibernateUtil.close();
 
-            return newUser;
+                return newUser;
 
-        }catch (HibernateException e){
-            HibernateUtil.rollback();
-            throw e;
+            } catch (HibernateException e) {
+                HibernateUtil.rollback();
+                HibernateUtil.close();
+                throw e;
+            }
         }
     }
 
@@ -83,30 +103,44 @@ public class UserDao implements IUserDao {
      * @return new User
      * @throws UserAlreadyExistException if user which this email already exist
      */
-    public User createUser(String email, String password, UserAccessGroup userAccessGroup, Date birthDate) throws UserAlreadyExistException {
+    public User createUser(String email, String password, UserAccessGroup userAccessGroup, Date birthDate) throws UserAlreadyExistException,
+            IllegalArgumentException {
 
-        try {
-            User currentUser = getUser(email);
-            if (currentUser != null) {
-                throw new UserAlreadyExistException("User with email: " + email + " is already exist.");
+        if(email==null) {
+            throw new IllegalArgumentException("argument \"email\": value " + email + " is not valid");
+        } else if (password==null) {
+            throw new IllegalArgumentException("argument \"password\": value " + password + " is not valid");
+        } else if (userAccessGroup==null) {
+            throw new IllegalArgumentException("argument \"userAccessGroup\": value " + userAccessGroup + " is not valid");
+        } else if (birthDate==null) {
+            throw new IllegalArgumentException("argument \"birthDate\": value " + birthDate + " is not valid");
+        } else {
+
+            try {
+                User currentUser = getUser(email);
+                if (currentUser != null) {
+                    throw new UserAlreadyExistException("User with email: " + email + " is already exist.");
+                }
+            } catch (UserNotFoundException e) {
+                // ok - create a new user
             }
-        }catch (UserNotFoundException e){
-            // ok - create a new user
-        }
 
-        try {
+            try {
 
-            HibernateUtil.begin();
-            Session session = HibernateUtil.getSession();
-            User newUser = new User(email, password, birthDate, userAccessGroup);
-            session.save(newUser);
-            HibernateUtil.commit();
+                HibernateUtil.begin();
+                Session session = HibernateUtil.getSession();
+                User newUser = new User(email, password, birthDate, userAccessGroup);
+                session.save(newUser);
+                HibernateUtil.commit();
+                HibernateUtil.close();
 
-            return newUser;
+                return newUser;
 
-        }catch (HibernateException e){
-            HibernateUtil.rollback();
-            throw e;
+            } catch (HibernateException e) {
+                HibernateUtil.rollback();
+                HibernateUtil.close();
+                throw e;
+            }
         }
     }
 
@@ -115,28 +149,34 @@ public class UserDao implements IUserDao {
      * @param user user which need to delete
      * @throws UserNotFoundException if this user doesn't exist in DataBase
      */
-    public  void deleteUser(User user) throws UserNotFoundException {
+    public  void deleteUser(User user) throws UserNotFoundException, IllegalArgumentException {
 
-        try {
-            User currentUser = getUser(user.getEmail());
+        if(user!=null) {
 
             try {
+                User currentUser = getUser(user.getEmail());
 
-                HibernateUtil.begin();
-                Session session = HibernateUtil.getSession();
-                session.delete(currentUser);
-                HibernateUtil.commit();
+                try {
 
-            }catch (HibernateException e){
-                HibernateUtil.rollback();
+                    HibernateUtil.begin();
+                    Session session = HibernateUtil.getSession();
+                    session.delete(currentUser);
+                    HibernateUtil.commit();
+                    HibernateUtil.close();
+
+                } catch (HibernateException e) {
+                    HibernateUtil.rollback();
+                    HibernateUtil.close();
+                    throw e;
+                }
+
+            } catch (UserNotFoundException e) {
+                // user not found - throw exception
                 throw e;
             }
-
-        }catch (UserNotFoundException e){
-            // user not found - throw exception
-            throw e;
+        } else {
+            throw new UserNotFoundException("User: " + user + " was not found");
         }
-
     }
 
     /**
@@ -147,18 +187,25 @@ public class UserDao implements IUserDao {
      */
     public UserAccessGroup getUserAccessGroup(String name) throws UserAccessGroupNotFoundException {
 
-        HibernateUtil.begin();
-        Session session = HibernateUtil.getSession();
-        Query query = session.createQuery("from UserAccessGroup where name = :name");
-        query.setParameter("name", name);
-        UserAccessGroup currentGroup = (UserAccessGroup)query.uniqueResult();
-        HibernateUtil.commit();
+        if(name!=null) {
 
-        if(currentGroup==null){
+            HibernateUtil.begin();
+            Session session = HibernateUtil.getSession();
+            Query query = session.createQuery("from UserAccessGroup where name = :name");
+            query.setParameter("name", name);
+            UserAccessGroup currentGroup = (UserAccessGroup) query.uniqueResult();
+            HibernateUtil.commit();
+            HibernateUtil.close();
+
+            if (currentGroup == null) {
+                throw new UserAccessGroupNotFoundException("User access group with name: " + name + " not found.");
+            }
+
+            return currentGroup;
+
+        } else {
             throw new UserAccessGroupNotFoundException("User access group with name: " + name + " not found.");
         }
-
-        return currentGroup;
 
     }
 
@@ -167,34 +214,41 @@ public class UserDao implements IUserDao {
      * @param name user access group name
      * @param isAdmin if admin - true, else - false
      * @return new user access group
-     * @throws UserAccessGroupAlreadyExistException if user access group which this name already exist
+     * @throws UserAccessGroupAlreadyExistException if user access group which this name already exists
      */
-    public UserAccessGroup createUserAccessGroup(String name, boolean isAdmin) throws UserAccessGroupAlreadyExistException{
+    public UserAccessGroup createUserAccessGroup(String name, boolean isAdmin) throws UserAccessGroupAlreadyExistException,
+            IllegalArgumentException {
 
-        try {
-            UserAccessGroup currentGroup = getUserAccessGroup(name);
-            if (currentGroup != null) {
-                throw new UserAccessGroupAlreadyExistException("User access group with name: " + name + " is already exist.");
+        if (name != null) {
+
+            try {
+                UserAccessGroup currentGroup = getUserAccessGroup(name);
+                if (currentGroup != null) {
+                    throw new UserAccessGroupAlreadyExistException("User access group with name: " + name + " is already exist.");
+                }
+            } catch (UserAccessGroupNotFoundException e) {
+                // ok - create a new user access group
             }
-        }catch (UserAccessGroupNotFoundException e){
-            // ok - create a new user access group
+
+            try {
+
+                HibernateUtil.begin();
+                Session session = HibernateUtil.getSession();
+                UserAccessGroup newGroup = new UserAccessGroup(name, isAdmin);
+                session.save(newGroup);
+                HibernateUtil.commit();
+                HibernateUtil.close();
+
+                return newGroup;
+
+            } catch (HibernateException e) {
+                HibernateUtil.rollback();
+                HibernateUtil.close();
+                throw e;
+            }
+        } else {
+            throw new IllegalArgumentException("argument \"name\": value " + name + " is not valid");
         }
-
-        try {
-
-            HibernateUtil.begin();
-            Session session = HibernateUtil.getSession();
-            UserAccessGroup newGroup = new UserAccessGroup(name, isAdmin);
-            session.save(newGroup);
-            HibernateUtil.commit();
-
-            return newGroup;
-
-        }catch (HibernateException e){
-            HibernateUtil.rollback();
-            throw e;
-        }
-
     }
 
     /**
@@ -202,26 +256,33 @@ public class UserDao implements IUserDao {
      * @param userAccessGroup user access group which need to delete
      * @throws UserAccessGroupNotFoundException if this user access group doesn't exist in DataBase
      */
-    public void deleteUserAccessGroup(UserAccessGroup userAccessGroup) throws UserAccessGroupNotFoundException{
+    public void deleteUserAccessGroup(UserAccessGroup userAccessGroup) throws UserAccessGroupNotFoundException,
+            IllegalArgumentException {
 
-        try {
-            UserAccessGroup currentGroup = getUserAccessGroup(userAccessGroup.getName());
-
+        if(userAccessGroup!=null) {
             try {
+                UserAccessGroup currentGroup = getUserAccessGroup(userAccessGroup.getName());
 
-                HibernateUtil.begin();
-                Session session = HibernateUtil.getSession();
-                session.delete(currentGroup);
-                HibernateUtil.commit();
+                try {
 
-            }catch (HibernateException e){
-                HibernateUtil.rollback();
+                    HibernateUtil.begin();
+                    Session session = HibernateUtil.getSession();
+                    session.delete(currentGroup);
+                    HibernateUtil.commit();
+                    HibernateUtil.close();
+
+                } catch (HibernateException e) {
+                    HibernateUtil.rollback();
+                    HibernateUtil.close();
+                    throw e;
+                }
+
+            } catch (UserAccessGroupNotFoundException e) {
+                // user not found - throw exception
                 throw e;
             }
-
-        }catch (UserAccessGroupNotFoundException e){
-            // user not found - throw exception
-            throw e;
+        } else {
+            throw new UserAccessGroupNotFoundException("User access group: " + userAccessGroup + " was not found");
         }
     }
